feat(): abstract useTbaAddress hook
